draft for edit-review done, last presend commit

+++ routes/edit.js
@@ -18,16 +18,40 @@ const storage = multer.diskStorage({
 
 const upload = multer({ storage: storage })
 
-router.get('/', checkAuthenticate, (req, res) => {
-    const user = req.user
-
-    if (!user) {
+router.get('/user', checkAuthenticate, (req, res) => {
+    if (!req.isAuthenticated()) {
         res.redirect("/error?errorMsg=Login details could not be found.")
         return
     }
 
-    console.log(`ROUTE -> EDIT PROFILE (${user.name})`)
-    res.render("edit", user)
+    console.log(`ROUTE -> EDIT PROFILE (${req.user.name})`)
+    res.render("edit-profile", req.user)
+})
+
+router.get("/review/:revId", checkAuthenticate, async (req, res) => {
+    try {
+        if (!req.isAuthenticated()) {
+            error.throwLoginError()
+        }
+
+        const review = await query.getReview({ _id: req.params.revId })
+
+        if (!review) {
+            error.throwReviewFetchError()
+        }
+
+        console.log(`ROUTE -> EDIT REVIEW (${review.title})`)
+        console.log(review)
+        res.render("edit-review", review)
+    } catch (err) {
+        console.log(`ERROR! ${err.message}`)
+
+        if (err.name != "ReviewFetchError") {
+            res.redirect(`/error`)
+        } else {
+            res.redirect(`/error?errorMsg=${err.message}`)
+        }
+    }
 })
 
 router.post('/update', upload.single("avatar"), async (req, res) => {
@@ -48,13 +72,18 @@ router.post('/update', upload.single("avatar"), async (req, res) => {
             const updateObj = { name: name, description: desc }
             if (avatar) {
                 updateObj.avatar = avatar.filename
-                fs.unlink("./public/imgs/avatars/" + oldAvatar, (err) => {
-                    if (err) {
-                        console.log(err)
-                    } else {
-                        console.log("Deleted old avatar of " + user.name)
-                    }
-                })
+
+                if (oldAvatar !== "default_avatar.png") {
+                    fs.unlink("./public/imgs/avatars/" + oldAvatar, (err) => {
+                        if (err) {
+                            console.log(err)
+                        } else {
+                            console.log("Deleted old avatar of " + user.name)
+                        }
+                    })
+                } else {
+                    console.log("Switched old avatar of " + user.name)
+                }
             }
 
             await query.updateProfile({ _id: user._id }, { $set: updateObj })
